Refactorizacion de carrito de compras

+++ src/controllers/carts.controllers.js
@@ -1,35 +1,58 @@
 import { request, response } from "express";
-import { CartsRepository } from "../repositories/index.js";
+import {
+  CartsRepository,
+  ProductsRepository,
+  UsersRepository,
+} from "../repositories/index.js";
 
 export const getCartById = async (req = request, res = response) => {
   try {
+    const { _id } = req;
     const { cid } = req.params;
-    // const carrito = await getCartByIdService(cid);
+
+    const usuario = await UsersRepository.getUserById(_id);
+
+    if (!usuario)
+      return res.status(400).json({ ok: false, msg: "Usuario no existe" });
+
+    if (!(usuario.cart_id, toString() === cid))
+      return res.status(400).json({ ok: false, msg: "Carrito no valido" });
+
     const carrito = await CartsRepository.getCartById(cid);
 
-    if (carrito) return res.json({ carrito });
-
-    return res.status(404).json({ msg: `El carrito con id ${cid} no existe` });
+    return res.json({ carrito });
   } catch (error) {
     return res.status(500).json({ msg: "Hablar con un administrador" });
   }
 };
 
-export const createCart = async (req = request, res = response) => {
-  try {
-    // const carrito = await createCartService();
-    const carrito = await CartsRepository.createCart();
-    return res.json({ msg: "Carrito creado", carrito });
-  } catch (error) {
-    console.log("createCart ->", error);
-    return res.status(500).json({ msg: "Hablar con un administrador" });
-  }
-};
+// export const createCart = async (req = request, res = response) => {
+//   try {
+//     const carrito = await CartsRepository.createCart();
+//     return res.json({ msg: "Carrito creado", carrito });
+//   } catch (error) {
+//     console.log("createCart ->", error);
+//     return res.status(500).json({ msg: "Hablar con un administrador" });
+//   }
+// };
 
 export const addProductInCart = async (req = request, res = response) => {
   try {
+    // _id usuario logueado(token)
+    const { _id } = req;
     const { cid, pid } = req.params;
-    // const carrito = await addProductInCartService(cid, pid);
+
+    const usuario = await UsersRepository.getUserById(_id);
+    if (!usuario)
+      return res.status(400).json({ ok: false, msg: "Usuario no existe" });
+
+    if (!(usuario.cart_id, toString() === cid))
+      return res.status(400).json({ ok: false, msg: "Carrito no valido" });
+
+    const existeProducto = await ProductsRepository.getProductById(pid);
+    if (!existeProducto)
+      return res.status(400).json({ ok: false, msg: "El producto no existe!" });
+
     const carrito = await CartsRepository.addProductInCart(cid, pid);
 
     if (!carrito)
@@ -45,11 +68,23 @@ export const addProductInCart = async (req = request, res = response) => {
 
 export const deleteProductsInCart = async (req = request, res = response) => {
   try {
+    const { _id } = req;
     const { cid, pid } = req.params;
-    // const carrito = await deleteProductsInCartService(cid, pid);
+
+    const usuario = await UsersRepository.getUserById(_id);
+
+    if (!usuario)
+      return res.status(400).json({ ok: false, msg: "Usuario no existe" });
+
+    if (!(usuario.cart_id, toString() === cid))
+      return res.status(400).json({ ok: false, msg: "Carrito no valido" });
+
+    const existeProducto = await ProductsRepository.getProductById(pid);
+    if (!existeProducto)
+      return res.status(400).json({ ok: false, msg: "El producto no existe!" });
+
     const carrito = await CartsRepository.deleteProductsInCart(cid, pid);
-    if (!carrito)
-      return res.status(404).json({ msg: "No se pudo realizar la operacion" });
+
     return res.json({ msg: "Producto eliminado del carrito" });
   } catch (error) {
     return res
@@ -60,8 +95,21 @@ export const deleteProductsInCart = async (req = request, res = response) => {
 
 export const updateProductsInCart = async (req = request, res = response) => {
   try {
+    const { _id } = req;
     const { cid, pid } = req.params;
     const { quantity } = req.body;
+
+    const usuario = await UsersRepository.getUserById(_id);
+
+    if (!usuario)
+      return res.status(400).json({ ok: false, msg: "Usuario no existe" });
+
+    if (!(usuario.cart_id, toString() === cid))
+      return res.status(400).json({ ok: false, msg: "Carrito no valido" });
+
+    const existeProducto = await ProductsRepository.getProductById(pid);
+    if (!existeProducto)
+      return res.status(400).json({ ok: false, msg: "El producto no existe!" });
 
     if (!quantity || !Number.isInteger(quantity))
       return res.status(404).json({
@@ -83,17 +131,17 @@ export const updateProductsInCart = async (req = request, res = response) => {
   }
 };
 
-export const deleteCart = async (req = request, res = response) => {
-  try {
-    const { cid } = req.params;
+// export const deleteCart = async (req = request, res = response) => {
+//   try {
+//     const { cid } = req.params;
 
-    // const carrito = await deleteCartService(cid);
-    const carrito = await CartsRepository.deleteCart(cid);
+//     // const carrito = await deleteCartService(cid);
+//     const carrito = await CartsRepository.deleteCart(cid);
 
-    if (!carrito)
-      return res.status(404).json({ msg: "No se pudo realizar la operacion" });
-    return res.json({ msg: "Producto actualizado en el carrito", carrito });
-  } catch (error) {
-    return res.status(500).json({ msg: "Hablar con un administrador" });
-  }
-};
+//     if (!carrito)
+//       return res.status(404).json({ msg: "No se pudo realizar la operacion" });
+//     return res.json({ msg: "Producto actualizado en el carrito", carrito });
+//   } catch (error) {
+//     return res.status(500).json({ msg: "Hablar con un administrador" });
+//   }
+// };
